move xtstatus und readproof to one level above
