+++ src/std/rpc/mod.rs
@@ -14,8 +14,6 @@
    limitations under the License.
 
 */
-use serde::{Deserialize, Serialize};
-
 #[cfg(feature = "ws-client")]
 pub use ws_client::WsRpcClient;
 
@@ -36,28 +34,4 @@ pub enum RpcClientError {
 	NoErrorInformationFound(String),
 	#[error(transparent)]
 	Other(#[from] Box<dyn std::error::Error + Send + Sync + 'static>),
-}
-
-#[derive(Debug, PartialEq, Eq, Copy, Clone)]
-pub enum XtStatus {
-	Unknown = 0,
-	/// uses `author_submit` without watching.
-	SubmitOnly = 1,
-	Ready = 2,
-	Broadcast = 3,
-	InBlock = 4,
-	Finalized = 5,
-	Future = 10,
-}
-
-// Exact structure from
-// https://github.com/paritytech/substrate/blob/master/client/rpc-api/src/state/helpers.rs
-// Adding manually so we don't need sc-rpc-api, which brings in async dependencies
-#[derive(Debug, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct ReadProof<Hash> {
-	/// Block hash used to generate the proof
-	pub at: Hash,
-	/// A proof used to prove that storage entries are included in the storage trie
-	pub proof: Vec<sp_core::Bytes>,
 }
